Improve city banner tooltip without overriding original file
1. Use decorator pattern to update the city banner food growth tooltip when there's an update to the current food or the population. This way you won't need to override the original city-banner.js file for better compatibility with other mods.
2. Also added a new check to only show this tooltip when hovering over current player's city.
